requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
APScheduler==3.10.4
librouteros==3.2.1
pytz==2023.3
python-dateutil==2.8.2
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
six==1.16.0
tzlocal==5.2
